move ALS close test to its own file to avoid node-http.test.ts's pre-existing proxy flake
